// src/types/transit/oba.ts
export interface ObaRouteSearchResult {
  id: string;
  shortName: string;
  longName?: string;
  description?: string;
  agencyId: string;
  agencyName?: string;
  url?: string;
  color?: string;
  textColor?: string;
  type?: number;
}

export interface ObaStopSearchResult {
  id: string;
  name: string;
  code: string;
  direction?: string;
  latitude: number;
  longitude: number;
  routeIds: string[];
  routes?: ObaRouteSearchResult[];
  wheelchairBoarding?: string;
  locationType?: number;
}